% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{oi_get}
\alias{oi_get}
\title{Get OSM data associated with a particular type of infrastructure}
\usage{
oi_get(place, infra_type = NULL, query = NULL, extra_tags = et, ...)
}
\arguments{
\item{place}{Description of the geographical area that should be matched
with a .osm.pbf file}

\item{infra_type}{Type of infrastructure, e.g. cycle_infrastructure}

\item{query}{A character string representing an SQL query to send to
\code{oe_get()}}

\item{extra_tags}{Which additional columns, corresponding to OSM tags,
should be in the resulting dataset?}

\item{...}{Arguments passed to \code{oe_get()}}
}
\description{
Get OSM data associated with a particular type of infrastructure
}
\examples{
highways = oi_get("isle of wight", infra_type = "highway")
cycleways = oi_get("isle of wight", infra_type = "cycle_infrastructure")
plot(cycleways)
cycleways_minimal = oi_get("isle of wight", infra_type = "cycle_infrastructure_minimal")
}
